turn on signing during the build
